feature: Move from_braket_instruction() to operation_helpers.py
from_braket_instruction() converts IR of gates and noises to GateOperation and KrausOperation. It makes more sense to move it to operation_helpers.py.

Add check_CPTP() in operation_helpers.py
